+++ code/js/src/main/components/states/nextStates/NextStateItem.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useEffect, useState } from 'react'
 import { Link } from 'react-router-dom'
 import { getSirenAction } from '../../../api/apiRoutes'
 import { deleteNextState } from '../../../api/states'
@@ -6,25 +6,41 @@ import { Credentials } from '../../../utils/userSession'
 
 type IssueStateProps = {
     state: IssueState
-    nextState: IssueState,
-    onRemove: () => void
+    nextState: IssueState
     onFinishRemove: (success: boolean, message: string) => void
     credentials: Credentials
 }
 
-function NextStateItem({ state, nextState, onRemove, onFinishRemove, credentials }: IssueStateProps): JSX.Element {
-    function removeState() {
-        onRemove()
+function NextStateItem({ state, nextState, onFinishRemove, credentials }: IssueStateProps): JSX.Element {
+    const [message, setMessage] = useState(null)
+    const [toRemove, setToRemove] = useState(false)
+    
+    useEffect(() => {
+        let isCancelled = false
 
-        deleteNextState(state.projectId, state.number, nextState.number, credentials)
-            .then(() => onFinishRemove(true, null))
-            .catch(err => onFinishRemove(false, err.message))
-    }
+        if (toRemove) {
+            setMessage('Removing state...')
+            deleteNextState(state.projectId, state.number, nextState.number, credentials)
+                .then(() => {
+                    if (isCancelled) return
+                    onFinishRemove(true, null)
+                })
+                .catch(err => {
+                    if (isCancelled) return
+                    onFinishRemove(false, err.message)
+                })
+        }
+
+        return () => {
+            isCancelled = true
+        }
+    }, [toRemove])
 
     return (
         <li>
             <p>Name: <Link to={`/projects/${state.projectId}/states/${nextState.number}`}>{nextState.name}</Link></p>
-            { getSirenAction(nextState.actions, 'delete-next-state') == null ? <></> : <button onClick={removeState}>Remove state</button> }
+            { getSirenAction(nextState.actions, 'delete-next-state') == null ? <></> : <button className="danger" onClick={() => setToRemove(true)}>Remove state</button> }
+            <p>{message}</p>
         </li>
     )
 }
